+++ test/performance_tests/calculus_performance_test.cpp
@@ -3,10 +3,10 @@
 
 static void BM_Addition(benchmark::State& state)
 {
-    std::vector<std::uint8_t> data1;
+    std::vector<std::uint16_t> data1;
     data1.resize(static_cast<std::size_t>(state.range(0)), 123u);
 
-    std::vector<std::uint8_t> data2;
+    std::vector<std::uint16_t> data2;
     data2.resize(static_cast<std::size_t>(state.range(1)), 123u);
     api::API const api_1 { data1 };
     api::API const api_2 { data2 };
@@ -17,10 +17,10 @@ static void BM_Addition(benchmark::State& state)
 
 static void BM_Subtraction(benchmark::State& state)
 {
-    std::vector<std::uint8_t> data1;
+    std::vector<std::uint16_t> data1;
     data1.resize(static_cast<std::size_t>(state.range(0)), 123u);
 
-    std::vector<std::uint8_t> data2;
+    std::vector<std::uint16_t> data2;
     data2.resize(static_cast<std::size_t>(state.range(1)), 123u);
     api::API const api_1 { data1 };
     api::API const api_2 { data2 };
@@ -31,10 +31,10 @@ static void BM_Subtraction(benchmark::State& state)
 
 static void BM_Multiplication(benchmark::State& state)
 {
-    std::vector<std::uint8_t> data1;
+    std::vector<std::uint16_t> data1;
     data1.resize(static_cast<std::size_t>(state.range(0)), 123u);
 
-    std::vector<std::uint8_t> data2;
+    std::vector<std::uint16_t> data2;
     data2.resize(static_cast<std::size_t>(state.range(1)), 123u);
     api::API const api_1 { data1 };
     api::API const api_2 { data2 };
@@ -45,10 +45,10 @@ static void BM_Multiplication(benchmark::State& state)
 
 static void BM_Division(benchmark::State& state)
 {
-    std::vector<std::uint8_t> data1;
+    std::vector<std::uint16_t> data1;
     data1.resize(static_cast<std::size_t>(state.range(0)), 123u);
 
-    std::vector<std::uint8_t> data2;
+    std::vector<std::uint16_t> data2;
     data2.resize(static_cast<std::size_t>(state.range(1)), 123u);
     api::API const api_1 { data1 };
     api::API const api_2 { data2 };
@@ -57,22 +57,23 @@ static void BM_Division(benchmark::State& state)
     }
 }
 
-BENCHMARK(BM_Addition)
+
+ BENCHMARK(BM_Addition)
     ->Unit(benchmark::kMillisecond)
     ->ArgsProduct({ benchmark::CreateRange(40000, 200000, /*multi=*/4),
         benchmark::CreateRange(40000, 200000, /*multi=*/4) });
 
-BENCHMARK(BM_Subtraction)
+ BENCHMARK(BM_Subtraction)
     ->Unit(benchmark::kMillisecond)
     ->ArgsProduct({ benchmark::CreateRange(40000, 200000, /*multi=*/4),
         benchmark::CreateRange(40000, 200000, /*multi=*/4) });
 
-BENCHMARK(BM_Multiplication)
+ BENCHMARK(BM_Multiplication)
     ->Unit(benchmark::kMillisecond)
     ->ArgsProduct({ benchmark::CreateRange(400, 2000, /*multi=*/4),
         benchmark::CreateRange(400, 2000, /*multi=*/4) });
 
-BENCHMARK(BM_Division)
+ BENCHMARK(BM_Division)
     ->Unit(benchmark::kMillisecond)
     ->ArgsProduct({ benchmark::CreateRange(400, 2000, /*multi=*/4),
         benchmark::CreateRange(400, 2000, /*multi=*/4) });
